Good enough to be @chadoh's local config
I've added in several vim shortcuts I like and simplified the prompt.

setup.sh is now AWESOME. It Just Works, on osx, and is very configurable
so it can be scripted to work on our servers.

One weird thing, with symlinking the gitconfig: since my .extra sets the
git user, it modifies .gitconfig. Since that's now symlinked, it ends up
being set in this project. We'll need to be careful to not accidentally
set that, but we probably will at some point, anyhow. Let's come up with
a better solution then.