Optimize for multi-servers & many users (.extra, .machine_nickname)
Check the readme for all the details.

.gitconfig is no longer a symlink, like all the other dotfiles, because
gitconfig ends up having user info in it. And we don't want anyone's
user info accidentally getting into this repo. Instead, you set that in
a ~/.extra file, which is sourced by .bashrc (and not included in the
repo).

Also, since a big focus of these dotfiles is being able to use them on
various servers as well as locally, it's nice to have a machine name in
the prompt. In my experience, the names of machines on ec2, and even my
own machine name, are not what I want to see in the prompt, though. When
I'm on my local machine, I want to see 'local'. When I'm on an ec2
instance, I likewise want some sort of nickname. This is accomplished by
adding a ~/.machine_nickname file, and reading its contents into the
prompt.

I also removed the username from the prompt, because how often are you
not your own user? Type `whoami` if you're ever confused. Including this
in the prompt seems like clutter to me.

Finally, I also tweaked the vimrc to look decent with light-background
color schemes.